week-5/day-1/daily/phone_finder_top/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-phonenumber-field==7.0.2
Faker==18.3.2
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
psycopg2==2.9.6
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
translate==3.6.1
tzdata==2023.3
urllib3==1.26.15
